✨ Feat: Add `ChildBuilder` widget
This commit introduces a new stateless widget called `ChildBuilder`.

The `ChildBuilder` widget:
- Takes a `child` widget and a `builder` function as arguments.
- The `builder` function receives the `BuildContext` and the `child` widget.
- Allows for constructing a widget tree where a part of it (the `child`) is passed down to be incorporated by the `builder` function, promoting widget composition and potentially optimization.

This is similar in concept to Flutter's built-in `Builder` widget, but specifically designed to pass an existing `child` widget instance to the `builder` function, which can be useful in scenarios like those handled by `AnimatedBuilder` or `TweenAnimationBuilder` where a part of the widget tree doesn't need to be rebuilt during animations.